worked on songs

+++ GentrysQuest.Game/Audio/Music/TimingPointsHandler.cs
@@ -4,9 +4,9 @@ namespace GentrysQuest.Game.Audio.Music
 {
     public class TimingPointsHandler
     {
-        private List<float> timingPoints;
+        private List<int> timingPoints;
 
-        public TimingPointsHandler(List<float> timingPoints)
+        public TimingPointsHandler(List<int> timingPoints)
         {
             this.timingPoints = timingPoints;
         }
